requirements.txt
flask
requests
sentence-transformers
numpy
endee
PyPDF2
googletrans==4.0.0-rc1
deep-translator
python-dotenv
beautifulsoup4
pytest
